Fixes #1: SQS Span Collector
An alternative Zipkin Span Collector implementation that leverages the AmazonSqsAsync interface to chain SQS reads.
This implementation also allows a list of queue URLs to pull messages from multiple sources.
Further this takes advantage of the thread pool provided by the Amazon client as well as prefetching capabilities of the buffered client.