Provides better fallbacks when client service name is missing
Arising from a conversation with Amazon X-Ray folks, the best behavior
when a remote service name is missing from a client span is:

If the intention is to not create a service link, don’t mark the
subsegment as remote

For the remote name, we can have a fallback model: start with the remote
service name, fallback to the hostname (if available), then the span
name (because it is already hinted as low cardinality). When there is a
problem, correct the remote service name.

Note: There is still a problem when there is double instrumentation,
because a remote will have a child of remote. Causing intermediate nodes
to show up.

See https://docs.google.com/document/d/1ZFg2SX2DUJFRE8MpodV9DL9uYgNHe-90PMQIb1DrY8k/edit
Fixes #88