license: removes copyright year and uses SPDX ID
As a small project, we have to conserve resources and not sign up for work that isn't required. I've recently realized many commercial and/or CNCF projects both use SPDX IDs and also don't bother with copyright year. If their legal team is ok with this, surely a volunteer team without access to one, should be, too! Doing so accomplishes the following:

* significantly increases readability of files, particularly small ones
* removes beginning of year maintenance, which cause a lot of FUD last year. IIRC some deployment failed and hours were spent in spite of docs.
* eliminates the need to do a full source check out, just to satisfy the license plugin. This means we can use actions/checkout defaults.

Same as https://github.com/openzipkin/zipkin-reporter-java/pull/257

Signed-off-by: Adrian Cole <adrian@tetrate.io>